requirements.txt
certifi==2024.2.2
charset-normalizer==3.3.2
idna==3.6
kafka==1.3.5
requests==2.31.0
urllib3==2.2.0
kafka-python==1.4.7
scikit-surprise
jproperties
pandas
flask
schedule
apscheduler
pytest
pytest-mock
coverage
aioredis
fastapi
uvicorn
dvc
flask_limiter
